Rename Map Services sample -> Search sample
