--- Android/samples/mapservices/app/build.gradle
@@ -1,45 +1,0 @@
-apply plugin: 'com.android.application'
-apply from: 'secrets.gradle'
-
-android {
-    compileSdkVersion 28
-    defaultConfig {
-        applicationId "com.microsoft.maps.sample.mapservices"
-        minSdkVersion 16
-        targetSdkVersion 28
-        versionCode 1
-        versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-        buildTypes.each {
-            it.buildConfigField "String", "CREDENTIALS_KEY", "\"$credentialsKey\""
-        }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-}
-
-repositories {
-    maven {
-        url  "https://microsoft-maps.bintray.com/Maven"
-    }
-}
-
-dependencies {
-    implementation fileTree(include: ['*.aar'], dir: 'libs')
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:design:28.0.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.android.volley:volley:1.1.1'
-    implementation 'com.microsoft.maps:maps-sdk:0.1.1'
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-}
