--- Android/samples/mapservices/app/src/main/res/values/strings.xml
@@ -1,5 +1,0 @@
-<resources>
-
-<string name="app_name">Bing Maps MapServices Sample</string>
-
-</resources>
