+++ Android/samples/sdksample/app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
 
-<string name="app_name">Bing Maps Sample App</string>
+<string name="app_name">Bing Maps SDK Sample</string>
 
 <string-array name="map_styles">
     <item>Symbolic Light</item>
