+++ Android/samples/search/README.md
