+++ Android/samples/search/app/proguard-rules.pro
