+++ Android/samples/search/app/src/main/res/drawable/button_active.xml
