+++ Android/samples/search/app/src/main/res/drawable/button_default.xml
