+++ Android/samples/search/app/src/main/res/drawable/ic_pin.xml
