+++ Android/samples/search/app/src/main/res/layout/activity_main.xml
