+++ Android/samples/search/app/src/main/res/layout/dialog_geocode.xml
