+++ Android/samples/search/app/src/main/res/layout/menu_demo.xml
