+++ Android/samples/search/app/src/main/res/mipmap-xxhdpi/ic_launcher.png
