+++ Android/samples/search/app/src/main/res/values/colors.xml
