+++ Android/samples/search/app/src/main/res/values/strings.xml
@@ -1,0 +1,5 @@
+<resources>
+
+<string name="app_name">Bing Maps Search Sample</string>
+
+</resources>
