+++ Android/samples/search/app/src/main/res/values/styles.xml
