+++ Android/samples/search/build.gradle
