+++ Android/samples/search/gradle.properties
