+++ Android/samples/search/gradle/wrapper/gradle-wrapper.properties
