+++ Android/samples/search/gradlew
