+++ Android/samples/search/gradlew.bat
