+++ Android/samples/search/settings.gradle
